added NewPost component to wrap the input creation and post creation
