+++ modules/HomePage/components/NewPost/NewPost.module.css
@@ -1,0 +1,7 @@
+.input-creation {
+  @apply shadow-soft p-4 bg-white rounded-md mb-6;
+}
+
+.input-placeholder {
+  @apply absolute h-full w-full z-10 cursor-pointer;
+}
