+++ modules/HomePage/pages/home-page.tsx
@@ -1,15 +1,11 @@
-import React, { useState } from 'react';
-import type { FC, ReactElement, MouseEvent } from 'react';
+import React from 'react';
+import type { FC, ReactElement } from 'react';
 import type { IPostFeed } from '@modules/shared/types/postFeed/IPostFeed';
 import Navigation from '@modules/shared/components/molecules/Navigation/Navigation';
 import Widget from '@modules/shared/components/atoms/Widget/Widget';
-import PostCreation from '@modules/shared/components/organisms/PostCreation/PostCreation';
-import Modal from '@modules/shared/components/organisms/Modal/Modal';
-import TextInput from '@modules/shared/components/atoms/TextInputs/TextInput';
-import * as ETextInput from '@modules/shared/components/atoms/TextInputs/types/ETextInput';
-import Avatar from '@modules/shared/components/atoms/Avatar/Avatar';
 import styles from './home-page.module.css';
 import Posts from '../components/Posts';
+import NewPost from '../components/NewPost/NewPost';
 
 interface WidgetData {
   content: string;
@@ -17,7 +13,6 @@ interface WidgetData {
 }
 
 const HomePage: FC<IPostFeed.IPosts> = ({ data }): ReactElement => {
-  const [showModal, setShowModal] = useState(false);
   const widgetData: WidgetData[] = [
     {
       id: '1',
@@ -36,50 +31,12 @@ const HomePage: FC<IPostFeed.IPosts> = ({ data }): ReactElement => {
     },
   ];
 
-  const showModalHandler = (): void => {
-    setShowModal(true);
-  };
-
-  const closeModalHandler = (e: MouseEvent<HTMLDivElement>): void => {
-    e.stopPropagation();
-    setShowModal(false);
-  };
-
   return (
     <section className="bg-grey-bg2 min-h-screen relative">
       <Navigation />
       <section className={styles['layout-parent']}>
         <div className={styles['posts-feed']}>
-          <div className={styles['input-creation']}>
-            <div className="flex">
-              <div className="mr-4">
-                <Avatar size="medium" variant="notFilled" />
-              </div>
-              <div className="relative flex w-full">
-                <div
-                  className="absolute h-full w-full z-10 cursor-pointer"
-                  onClick={showModalHandler}
-                  role="button"
-                  aria-hidden
-                />
-                <TextInput
-                  variants={ETextInput.Variants.Default}
-                  inputType={ETextInput.InputType.Default}
-                  id="id_1"
-                  value=""
-                  placeholder="What do you want to ask about?"
-                  onBlurInputHandler={(): boolean => true}
-                  onChangeInputValueHandler={(): boolean => true}
-                  onClickDeleteInputValueHandler={(): boolean => true}
-                />
-              </div>
-            </div>
-          </div>
-          {showModal && (
-            <Modal closeModalHandler={closeModalHandler}>
-              <PostCreation />
-            </Modal>
-          )}
+          <NewPost />
           <Posts data={data} />
         </div>
         <div className={styles.widgets}>
