+++ modules/shared/components/organisms/Modal/IModal.d.ts
@@ -1,8 +1,6 @@
-import type { MouseEvent } from 'react';
-
 declare namespace IModal {
   export interface IProps {
-    closeModalHandler: (e: MouseEvent<HTMLDivElement>) => void;
+    closeModalHandler: () => void;
   }
 }
