+++ modules/shared/components/organisms/PostCreation/PostCreation.stories.tsx
@@ -1,12 +1,15 @@
 import type { ReactElement } from 'react';
 import type { Story } from '@storybook/react';
 import PostCreation from './PostCreation';
+import type { IPostCreation } from './types/IPostCreation';
 
 export default {
   title: 'organisms/PostCreation',
   component: PostCreation,
 };
 
-const Template: Story = (args): ReactElement => <PostCreation {...args} />;
+const Template: Story<IPostCreation.IProps> = (args): ReactElement => (
+  <PostCreation {...args} />
+);
 export const Default = Template.bind({});
 Default.args = {};
